+++ User/bsp/a/bsp_spi.c
@@ -148,7 +148,7 @@ void HAL_SPI_AbortCpltCallback(SPI_HandleTypeDef *hspi) {
 }
 
 /* Exported functions ------------------------------------------------------- */
-int BSP_SPI_RegisterCallback(BSP_SPI_t spi, BSP_SPI_Callback_t type,
+int8_t BSP_SPI_RegisterCallback(BSP_SPI_t spi, BSP_SPI_Callback_t type,
                              void (*callback)(void)) {
   if (callback == NULL) return -1;
 
@@ -245,7 +245,7 @@ int BSP_SPI_RegisterCallback(BSP_SPI_t spi, BSP_SPI_Callback_t type,
   return 0;
 }
 
-int BSP_SPI_Transmit(BSP_SPI_t spi, uint8_t *data, uint16_t len) {
+int8_t BSP_SPI_Transmit(BSP_SPI_t spi, uint8_t *data, uint16_t len) {
   if (data == NULL) return -1;
 
   switch (spi) {
@@ -269,7 +269,7 @@ int BSP_SPI_Transmit(BSP_SPI_t spi, uint8_t *data, uint16_t len) {
   return 0;
 }
 
-int BSP_SPI_Receive(BSP_SPI_t spi, uint8_t *data, uint16_t len) {
+int8_t BSP_SPI_Receive(BSP_SPI_t spi, uint8_t *data, uint16_t len) {
   if (data == NULL) return -1;
 
   switch (spi) {
